+++ pyxem/tests/generators/test_indexation_generator.py
@@ -21,27 +21,20 @@ import numpy as np
 import hyperspy.api as hs
 
 from hyperspy._signals.signal2d import Signal2D
-
-from pyxem import ElectronDiffraction2D
-from pyxem.signals.indexation_results import TemplateMatchingResults
-from pyxem.generators.indexation_generator import (
-    IndexationGenerator,
-    TemplateIndexationGenerator,
-    ProfileIndexationGenerator,
-    VectorIndexationGenerator)
-
-
 from diffsims.libraries.vector_library import DiffractionVectorLibrary
-from diffsims.libraries.diffraction_library import DiffractionLibrary
 from diffsims.sims.diffraction_simulation import ProfileSimulation
-from pyxem.signals.diffraction_vectors import DiffractionVectors
-
-
 from diffsims.generators.diffraction_generator import DiffractionGenerator
 from diffsims.generators.library_generator import DiffractionLibraryGenerator
-from diffsims.libraries.diffraction_library import DiffractionLibrary
 from diffsims.libraries.structure_library import StructureLibrary
 
+from pyxem.generators import (
+    TemplateIndexationGenerator,
+    ProfileIndexationGenerator,
+    VectorIndexationGenerator
+)
+from pyxem.signals import (
+    ElectronDiffraction2D, TemplateMatchingResults, DiffractionVectors
+)
 from pyxem.utils.indexation_utils import OrientationResult
 
 def test_old_indexer_routine():