+++ pyxem/tests/signals/test_diffraction_variance1d.py
@@ -17,10 +17,9 @@
 # along with pyXem.  If not, see <http://www.gnu.org/licenses/>.
 
 import numpy as np
-import pytest
 
 from hyperspy.signals import Signal1D
-from pyxem.signals.diffraction_variance1d import DiffractionVariance1D
+from pyxem.signals import DiffractionVariance1D
 
 
 class TestDiffractionVariance1D: