+++ pyxem/tests/signals/test_indexation_results.py
@@ -18,16 +18,20 @@
 
 import numpy as np
 import pytest
-import diffsims as ds
 from matplotlib import pyplot as plt
-
-from pyxem.signals.indexation_results import TemplateMatchingResults, VectorMatchingResults
-from pyxem.signals.diffraction_vectors import DiffractionVectors
-from diffsims.libraries.vector_library import DiffractionVectorLibrary
-from diffsims.libraries.structure_library import StructureLibrary
-from pyxem.utils.indexation_utils import OrientationResult
-from pyxem.generators.indexation_generator import TemplateIndexationGenerator
 from transforms3d.euler import euler2mat
+
+from diffsims.libraries.structure_library import StructureLibrary
+from diffsims.generators.diffraction_generator import DiffractionGenerator
+from diffsims.generators.library_generator import DiffractionLibraryGenerator
+
+from pyxem.generators import TemplateIndexationGenerator
+from pyxem.signals import (
+    TemplateMatchingResults,
+    VectorMatchingResults,
+    DiffractionVectors,
+)
+from pyxem.utils.indexation_utils import OrientationResult
 
 
 def test_TemplateMatchingResults_to_crystal_map():
@@ -36,11 +40,11 @@ def test_TemplateMatchingResults_to_crystal_map():
 
 def test_TemplateMatchingResults_plot_best_results_on_signal(diffraction_pattern,default_structure):
     """ Coverage testing """
-    edc = ds.DiffractionGenerator(300, 5e-2)
+    edc = DiffractionGenerator(300)
     half_side_length = 4
     rot_list = [[0,1,0],[1,0,0]]
 
-    diff_gen = ds.DiffractionLibraryGenerator(edc)
+    diff_gen = DiffractionLibraryGenerator(edc)
     struc_lib = StructureLibrary(["A"], [default_structure], [rot_list])
     library = diff_gen.get_diffraction_library(
         struc_lib,