+++ pyxem/tests/signals/test_metadata.py
@@ -16,8 +16,9 @@
 # You should have received a copy of the GNU General Public License
 # along with pyXem.  If not, see <http://www.gnu.org/licenses/>.
 
-from pyxem.signals.diffraction2d import Diffraction2D
 import numpy as np
+
+from pyxem.signals import Diffraction2D
 
 
 class TestAdjustMetadata: